+++ apps/astarte_housekeeping/lib/astarte_housekeeping/engine.ex
@@ -22,16 +22,29 @@ defmodule Astarte.Housekeeping.Engine do
   alias Astarte.Housekeeping.Config
   alias Astarte.Housekeeping.Queries
 
-  def create_realm(realm, public_key_pem, replication_factor, opts \\ []) do
+  def create_realm(
+        realm,
+        public_key_pem,
+        replication_factor,
+        device_registration_limit,
+        opts \\ []
+      ) do
     _ =
       Logger.info(
         "Creating new realm.",
         tag: "create_realm",
         realm: realm,
-        replication_factor: replication_factor
+        replication_factor: replication_factor,
+        device_registration_limit: device_registration_limit
       )
 
-    Queries.create_realm(realm, public_key_pem, replication_factor, opts)
+    Queries.create_realm(
+      realm,
+      public_key_pem,
+      replication_factor,
+      device_registration_limit,
+      opts
+    )
   end
 
   def get_health do
@@ -86,12 +99,16 @@ defmodule Astarte.Housekeeping.Engine do
 
   @doc """
   Updates a given realm using the provided values map. Fails if realm does not exists or values are invalid.
-  At the moment, only updating the realm jwt_public_key_pem is supported.
+  At the moment, updates to jwt_public_key_pem and device_registration_limit are supported.
   Returns a tuple:
   - either {:ok, updated_realm} where updated_realm is a map describing the realm
   - or {:error, reason} where reason is an atom describing the error.
   """
-  @spec update_realm(String.t(), %{:jwt_public_key_pem => String.t(), optional(any()) => any()}) ::
+  @spec update_realm(String.t(), %{
+          :jwt_public_key_pem => String.t(),
+          :device_registration_limit => any(),
+          optional(any()) => any()
+        }) ::
           {:ok, map()}
           | {:error, :invalid_update_parameters}
           | {:error, :realm_not_found}
@@ -112,17 +129,14 @@ defmodule Astarte.Housekeeping.Engine do
 
   defp do_update_realm(realm_name, update_attrs) do
     %{
-      jwt_public_key_pem: new_jwt_public_key_pem
+      jwt_public_key_pem: new_jwt_public_key_pem,
+      device_registration_limit: new_limit
     } = update_attrs
 
     with realm when is_map(realm) <- Queries.get_realm(realm_name),
-         :ok <- update_jwt_public_key_pem(realm_name, new_jwt_public_key_pem) do
-      updated_realm =
-        if new_jwt_public_key_pem != nil do
-          Map.put(realm, :jwt_public_key_pem, new_jwt_public_key_pem)
-        else
-          realm
-        end
+         :ok <- update_jwt_public_key_pem(realm_name, new_jwt_public_key_pem),
+         :ok <- update_device_registration_limit(realm_name, new_limit) do
+      updated_realm = merge_realm_status(realm, update_attrs)
 
       _ = Logger.info("Successful update of realm #{realm_name}", tag: "realm_update_success")
 
@@ -148,6 +162,52 @@ defmodule Astarte.Housekeeping.Engine do
     end
   end
 
+  defp update_device_registration_limit(realm_name, new_limit) when is_integer(new_limit) do
+    _ = Logger.info("Updating device registration limit", tag: "update_device_registration_limit")
+    set_device_registration_limit(realm_name, new_limit)
+  end
+
+  defp update_device_registration_limit(realm_name, :remove_limit) do
+    _ = Logger.info("Removing device registration limit", tag: "remove_device_registration_limit")
+    delete_device_registration_limit(realm_name)
+  end
+
+  defp update_device_registration_limit(_realm_name, nil) do
+    :ok
+  end
+
+  defp delete_device_registration_limit(realm_name) do
+    case Queries.delete_device_registration_limit(realm_name) do
+      {:ok, %Xandra.Void{}} ->
+        :ok
+
+      {:error, reason} ->
+        _ =
+          Logger.warn(
+            "Cannot delete device registration limit, error #{inspect(reason)}",
+            tag: "delete_device_registration_limit"
+          )
+
+        {:error, :delete_device_registration_limit_fail}
+    end
+  end
+
+  defp set_device_registration_limit(realm_name, new_limit) when is_integer(new_limit) do
+    case Queries.set_device_registration_limit(realm_name, new_limit) do
+      {:ok, %Xandra.Void{}} ->
+        :ok
+
+      {:error, reason} ->
+        _ =
+          Logger.warn(
+            "Cannot set device registration limit, error #{inspect(reason)}",
+            tag: "set_device_registration_limit"
+          )
+
+        {:error, :set_device_registration_limit_fail}
+    end
+  end
+
   defp is_realm_update_valid?(update_attrs) do
     # TODO from ScyllaDB >= 5.3, replication can be altered
     update_valid? =
@@ -162,5 +222,21 @@ defmodule Astarte.Housekeeping.Engine do
     end
 
     update_valid?
+  end
+
+  defp merge_realm_status(old_realm, %{device_registration_limit: nil} = update_attrs) do
+    attrs = Map.delete(update_attrs, :device_registration_limit)
+    merge_realm_status(old_realm, attrs)
+  end
+
+  defp merge_realm_status(old_realm, %{device_registration_limit: :remove_limit} = update_attrs) do
+    attrs = Map.delete(update_attrs, :device_registration_limit)
+    merge_realm_status(%{old_realm | device_registration_limit: nil}, attrs)
+  end
+
+  defp merge_realm_status(old_realm, update_attrs) do
+    Map.merge(old_realm, update_attrs, fn _k, old_value, new_value ->
+      if new_value == nil, do: old_value, else: new_value
+    end)
   end
 end